@@ -1,11 +1,15 @@
-const sveltePreprocess = require('svelte-preprocess');
+import preprocess from 'svelte-preprocess';
 
-module.exports = {
-  preprocess: [
-    sveltePreprocess({
-      postcss: {
-        plugins: [require('autoprefixer')],
-      },
-    }),
-  ],
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
+	preprocess: preprocess(),
+
+	kit: {
+		// hydrate the <div id="svelte"> element in src/app.html
+		target: '#svelte'
+	}
 };
+
+export default config;
